Integer: move `compliment` method outside of integer<> class
the purpose of integer class is to bring attentiveness and
thoughtfulness into integral conversions (lack of which can lead to many
baffling errors). however, tifa also strives to be as zero-cost as
possible

the amount of code generated for integral class is massive - up to 104
(!) instantiations of every operator (for every possible tifa/primitive
integer type combination). and optimizers are not always 100% reliable

it would be nice to have a flag to make the types s8, u8, ..., s64 just
aliases for primitive types sint8_t, uint8_t, ..., sint64_t; in fact,
this flag should be enabled by default in release builds

to make that happen, we have to move all the methods out of the
`integer` class so they can operate nicely also with primitive types